fixed bug in rgb with uint and in find_best with double

+++ sheet01/identification-Q234/find_best_match.m
@@ -24,7 +24,7 @@ function [best_match, D] = find_best_match(model_images, query_images, dist_type
 
   % compute distance matrix
   for i=1:lq
-     img1 = imread(char(query_images(i)));
+     img1 = double(imread(char(query_images(i))));
      
      % if hist is expecting a gray picture we convert it
      if (hist_isgray == 1)
@@ -34,7 +34,7 @@ function [best_match, D] = find_best_match(model_images, query_images, dist_type
      end 
      h1 = hist_func(img1, num_bins);
      for j=1:lm
-        img2 = imread(char(model_images(j)));
+        img2 = double(imread(char(model_images(j))));
         if (hist_isgray == 1) 
             if (size(img2,3) == 3)
                 img2 = rgb2gray(img2);
